refactor: useQuery 커스텀훅으로 만들어 컴포넌트와 분리 #113
